+++ frontend/src/pages/orders/OrderListPage.css
@@ -42,12 +42,23 @@
     font-weight: 700;
 }
 
-/* 주문서와 주문 목록을 Thymeleaf orders.html처럼 2열로 배치함 */
-.order-layout {
+/* 주문서 화면은 상품 확인과 수령 정보를 나란히 배치함 */
+.checkout-layout {
     display: grid;
     grid-template-columns: 0.92fr 1.08fr;
     gap: 22px;
     align-items: start;
+}
+
+/* 내 주문 내역은 주문서 없이 목록만 넓게 보여주도록 단일 열로 맞춤 */
+.order-list-layout {
+    display: grid;
+    gap: 22px;
+}
+
+/* 주문서 패널은 같은 높이로 늘리지 않고 각 내용 높이에 맞게 유지함 */
+.checkout-panel {
+    align-self: start;
 }
 
 /* 주문 패널은 기존 상품/장바구니 카드 표면과 맞춤 */
@@ -366,8 +377,8 @@
 }
 
 @media (max-width: 900px) {
-    /* 태블릿 이하에서는 주문서와 주문 목록을 세로로 배치함 */
-    .order-layout {
+    /* 태블릿 이하에서는 주문 상품과 수령 정보를 세로로 배치함 */
+    .checkout-layout {
         grid-template-columns: 1fr;
     }
 }
